MagicCannon
Using a for loop to go though 1-100, made the mistake of putting the conditonal with the most complexity at the bottom of the if loop. Should always be at the top since if there is a condition that applies before it, it will never run and thus this is a bug.