Big push, completed part 1 C# basics
Completed the final boss battle Manticore, and have made a quite cool solution to the first problem inside part 2, OOP. I finally managed to use switch cases appropriately, along with an enum, to bring a quite; in my experience, sophisticated solution to Simulas Test, page 135. Im very proud of this as checking the solution online, i went above and beyond to make my program a bit more sophisticated, adding a unlocked value for the enum, which at first dosent seem like much, but on top of that using a switch case to check for the desired commands and setting up the conditional logic, especially during the commands Open and close, since now, which my solution, there are two instances that need to be checked before any of these commands can run appropriatley, closed and unlocked. While similiar, they are different, and thus my code reflects that.

Backtracking a bit, completing the Manticore was one of my first big programs, and it is a bit messy, and dosent change text colours, but it works. I have found new confidence in myself; being able to create and solve a problem from scratch, and using methods appropriatley and well to reach that result. Like i said, the code is a bit messy, but it works, and thats all i can hope for right now!

I plan to go back to it later, and refactoring the code to be sleeker and run better. Alas, i will keep working and studying, as this feels great.